+++ src/raps/unlockAndCrosschainSwap.ts
@@ -1,70 +1,8 @@
 import { Address } from 'viem';
 
-import { add } from '@/helpers/utilities';
-
-import { assetNeedsUnlocking, estimateApprove } from './actions';
-import { estimateCrosschainSwapGasLimit } from './actions/crosschainSwap';
+import { assetNeedsUnlocking } from './actions';
 import { createNewAction, createNewRap } from './common';
 import { RapAction, RapSwapActionParameters, RapUnlockActionParameters } from './references';
-
-export const estimateUnlockAndCrosschainSwap = async ({
-  sellAmount,
-  quote,
-  chainId,
-  assetToSell,
-}: Pick<RapSwapActionParameters<'crosschainSwap'>, 'sellAmount' | 'quote' | 'chainId' | 'assetToSell'>) => {
-  const {
-    from: accountAddress,
-    sellTokenAddress,
-    allowanceTarget,
-    allowanceNeeded,
-  } = quote as {
-    from: Address;
-    sellTokenAddress: Address;
-    allowanceTarget: Address;
-    allowanceNeeded: boolean;
-  };
-
-  let gasLimits: (string | number)[] = [];
-  let swapAssetNeedsUnlocking = false;
-
-  if (allowanceNeeded) {
-    swapAssetNeedsUnlocking = await assetNeedsUnlocking({
-      owner: accountAddress,
-      amount: sellAmount,
-      assetToUnlock: assetToSell,
-      spender: allowanceTarget,
-      chainId,
-    });
-  }
-
-  if (swapAssetNeedsUnlocking) {
-    const unlockGasLimit = await estimateApprove({
-      owner: accountAddress,
-      tokenAddress: sellTokenAddress,
-      spender: allowanceTarget,
-      chainId,
-    });
-    gasLimits = gasLimits.concat(unlockGasLimit);
-  }
-
-  const swapGasLimit = await estimateCrosschainSwapGasLimit({
-    chainId,
-    requiresApprove: swapAssetNeedsUnlocking,
-    quote,
-  });
-
-  if (swapGasLimit === null || swapGasLimit === undefined || isNaN(Number(swapGasLimit))) {
-    return null;
-  }
-
-  const gasLimit = gasLimits.concat(swapGasLimit).reduce((acc, limit) => add(acc, limit), '0');
-  if (isNaN(Number(gasLimit))) {
-    return null;
-  }
-
-  return gasLimit.toString();
-};
 
 export const createUnlockAndCrosschainSwapRap = async (swapParameters: RapSwapActionParameters<'crosschainSwap'>) => {
   let actions: RapAction<'crosschainSwap' | 'unlock'>[] = [];
